+++ inc/refcounted.h
@@ -62,7 +62,7 @@ public:
 	// copy creation fcn
 	static RefCounted<T>* Create( T & v ) { return new RefCounted<T>( v ); }
 	// create with 'n' empty items
-	static RefCounted<T>* Create( int n ) { return new RefCounted<T>( n ); }
+	static RefCounted<T>* Create( size_t n ) { return new RefCounted<T>( n ); }
 	// 'slice' creation fcn
 	// TODO: inc ref the copied items!
 	static RefCounted<T>* Create( T & v, size_t start, size_t end ) { return new RefCounted<T>( v, start, end ); }
